move H to h.py
